refactor: implement `try_process_own_join_commit()` on `PendingConversation`
